requirements.txt
python-telegram-bot==12.7
python-dotenv
sqlite3